third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/computed/get-expected.txt

This is a testharness.js-based test.
[FAIL] Getting a custom property not in the computed style returns undefined
  assert_equals: expected (undefined) undefined but got (object) null
Harness: the test ran to completion.
